Moved PageService tests into the new JekyllItemService test class
